requirements.txt
python-dotenv~=1.0.1
kafka-python~=2.0.2